+++ packages/widgetbook_generator/lib/metadata.dart
@@ -1,2 +1,2 @@
 /// The version as in pubspec.yaml
-const packageVersion = '4.0.0-alpha.1';
+const packageVersion = '4.0.0-alpha.2';